pyproject.toml
[project]
name = "expensetracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "httpie>=3.2.4",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-utils>=0.42.0",
]
